switched to double values for node

+++ src/main/java/Main.java
@@ -8,7 +8,6 @@
  -----------------------------------------------------------------------------*/
 
 import Engine.Node;
-import org.w3c.dom.NodeList;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -64,7 +63,8 @@ public class Main {
                 if (!closedSet.contains(neighbor)) {
                     double tempG = gScores.get(current) +
                             getDistance(current, neighbor);
-                    if (!openSet.contains(neighbor) || gScores.get(neighbor) == null) {
+                    if (!openSet.contains(neighbor) ||
+                            gScores.get(neighbor) == null) {
                         cameFrom.put(neighbor, current);
                         gScores.put(neighbor, tempG);
                     } else if (tempG <= gScores.get(neighbor)) {
@@ -86,8 +86,8 @@ public class Main {
     }
 
     private static double getHeuristic(Node node) {
-        int a = end.getX() - node.getX();
-        int b = end.getY() - node.getY();
+        double a = end.getX() - node.getX();
+        double b = end.getY() - node.getY();
         double cSq = Math.pow(a, 2) + Math.pow(b, 2);
         return Math.sqrt(cSq);
     }
@@ -97,20 +97,20 @@ public class Main {
     }
     
     private static double getDistance(Node n1, Node n2) {
-        int a = n2.getX() - n1.getX();
-        int b = n2.getY() - n1.getY();
+        double a = n2.getX() - n1.getX();
+        double b = n2.getY() - n1.getY();
         double cSq = Math.pow(a, 2) + Math.pow(b, 2);
         return Math.sqrt(cSq);
     }
     
     private static ArrayList<Node> getNeighbors(Node n) {
         ArrayList<Node> neighbors = new ArrayList();
-        int x = n.getX();
-        int y = n.getY();
+        double x = n.getX();
+        double y = n.getY();
         for (int i = -1; i <= 1; i ++) {
             for (int j = -1; j <= 1; j ++) {
-                int neighborX = x + i;
-                int neighborY = y + j;
+                double neighborX = x + i;
+                double neighborY = y + j;
                 if ((neighborX >= 0 && neighborY >= 0) &&
                         !(neighborX == x && neighborY == y))  {
                     neighbors.add(new Node(neighborX, neighborY));
